add DataSaver interface and convert Saving-Methods with extract into XMLDataSaver class
