+++ src/Properties/AssemblyInfo.cs
@@ -7,4 +7,4 @@ using System.Runtime.InteropServices;
 
 [assembly: AssemblyVersion("1.0.2")]
 [assembly: AssemblyFileVersion("1.0.2")]
-[assembly: AssemblyInformationalVersion("1.0.2-1-g090b867")]
+[assembly: AssemblyInformationalVersion("1.0.2-2-g6cca6d0")]
